IGraphics WIN: fix mouse-wheel delta truncation
GET_WHEEL_DELTA_WPARAM (short) divided by WHEEL_DELTA (int 120) did integer
division before assigning to the float delta, so any sub-notch wheel delta
(< 120) truncated to 0. Precision touchpads stream sub-notch deltas, so fine
trackpad scrolling was dropped entirely and only full notches registered as
jumps. Cast both operands to float so fractional deltas pass through.

+++ IGraphics/Platforms/IGraphicsWin.cpp
@@ -492,7 +492,11 @@ LRESULT CALLBACK IGraphicsWin::WndProc(HWND hWnd, UINT msg, WPARAM wParam, LPARA
       else
       {
         IMouseInfo info = pGraphics->GetMouseInfo(lParam, wParam);
-        float d = GET_WHEEL_DELTA_WPARAM(wParam) / WHEEL_DELTA;
+        // Cast before dividing: GET_WHEEL_DELTA_WPARAM is a short and WHEEL_DELTA
+        // is 120, so the original int/int division truncated every sub-notch
+        // delta to 0. Precision touchpads stream deltas < 120, which were dropped
+        // entirely, making trackpad scrolling stutter (only full notches moved).
+        float d = static_cast<float>(GET_WHEEL_DELTA_WPARAM(wParam)) / static_cast<float>(WHEEL_DELTA);
         const float scale = pGraphics->GetTotalScale();
         RECT r;
         GetWindowRect(hWnd, &r);
